+++ client/src/pages/Performance.js
@@ -1,119 +1,78 @@
-<!DOCTYPE html>
-<html lang="ja">
-  <head>
-    <meta charset="UTF-8" />
-    <title>My To Dos!</title>
-    <link rel="stylesheet" href="css/styles.css" />
-    <script src="https://unpkg.com/react@16/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>
-    <script src="https://unpkg.com/babel-standalone@6.15.0/babel.min.js"></script>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="text/babel">
-      (() => {
-        const todos = [
-          { id: 0, title: "Task 0", isDone: false },
-          { id: 1, title: "Task 1", isDone: false },
-          { id: 2, title: "Task 2", isDone: true },
-        ];
+import React, { Component } from "react";
+import TodoItem from "../components/Todo/TodoItem";
 
-        function TodoItem(props) {
-          return (
-            <li key={props.todo.id}>
-              <label>
-                <input
-                  type="checkbox"
-                  checked={props.todo.isDone}
-                  onChange={() => props.checkTodo(props.todo)}
-                />
-                <span className={props.todo.isDone ? "done" : ""}>
-                  {props.todo.title}
-                </span>
-              </label>
-              <span
-                className="cmd"
-                onClick={() => props.deleteTodo(props.todo)}
-              >
-                [Finish!]
-              </span>
-            </li>
-          );
-        }
+class Performance extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      test: "",
+    };
+  }
 
-        function TodoList(props) {
-          const todos = props.todos.map((todo) => {
-            return (
-              <TodoItem
-                key={todo.id}
-                todo={todo}
-                checkTodo={props.checkTodo}
-                deleteTodo={props.deleteTodo}
-              />
-            );
-          });
-          return (
-            <ul>{props.todos.length ? todos : <li>Nothing to do!</li>}</ul>
-          );
-        }
+  render() {
+    return () => {
+      const todos = [
+        { id: 0, title: "Task 0", isDone: false },
+        { id: 1, title: "Task 1", isDone: false },
+        { id: 2, title: "Task 2", isDone: true },
+      ];
 
-        class App extends React.Component {
-          constructor() {
-            super();
-            this.state = {
-              todos: todos,
-            };
-            this.checkTodo = this.checkTodo.bind(this);
-            this.deleteTodo = this.deleteTodo.bind(this);
-          }
+    }
+}
+}}\
+// class Performance extends React.Component {
+//   constructor() {
+//     super();
+//     this.state = {
+//       todos: todos,
+//     };
+//     this.checkTodo = this.checkTodo.bind(this);
+//     this.deleteTodo = this.deleteTodo.bind(this);
+//   }
 
-          deleteTodo(todo) {
-            if (!confirm("Are you sure?")) {
-              return;
-            }
+//   deleteTodo(todo) {
+//     if (!confirm("Are you sure?")) {
+//       return;
+//     }
 
-            const todos = this.state.todos.slice();
-            const pos = this.state.todos.indexOf(todo);
+//     const todos = this.state.todos.slice();
+//     const pos = this.state.todos.indexOf(todo);
 
-            todos.splice(pos, 1);
-            this.setState({
-              todos: todos,
-            });
-          }
+//     todos.splice(pos, 1);
+//     this.setState({
+//       todos: todos,
+//     });
+//   }
 
-          checkTodo(todo) {
-            const todos = this.state.todos.map((todo) => {
-              return { id: todo.id, title: todo.title, isDone: todo.isDone };
-            });
+//   checkTodo(todo) {
+//     const todos = this.state.todos.map((todo) => {
+//       return { id: todo.id, title: todo.title, isDone: todo.isDone };
+//     });
 
-            const pos = this.state.todos
-              .map((todo) => {
-                return todo.id;
-              })
-              .indexOf(todo.id);
+//     const pos = this.state.todos
+//       .map((todo) => {
+//         return todo.id;
+//       })
+//       .indexOf(todo.id);
 
-            todos[pos].isDone = !todos[pos].isDone;
-            this.setState({
-              todos: todos,
-            });
-          }
+//     todos[pos].isDone = !todos[pos].isDone;
+//     this.setState({
+//       todos: todos,
+//     });
+//   }
 
-          render() {
-            return (
-              <div className="container">
-                <h1>My Todos</h1>
-                <TodoList
-                  todos={this.state.todos}
-                  checkTodo={this.checkTodo}
-                  deleteTodo={this.deleteTodo}
-                />
-              </div>
-            );
-          }
-        }
+//   render() {
+//     return (
+//       <div className="container">
+//         <h1>My Todos</h1>
+//         <TodoList
+//           todos={this.state.todos}
+//           checkTodo={this.checkTodo}
+//           deleteTodo={this.deleteTodo}
+//         />
+//       </div>
+//     );
+//   }
+// }
 
-        ReactDOM.render(<App />, document.getElementById("root"));
-      })();
-    </script>
-  </body>
-</html>
+export default Performance;
